fix(test) : implementacion save-file-test

+++ src/domain/use-cases/save-file.use-case.test.js
@@ -4,7 +4,8 @@ import { SaveFile } from './save-file.use-case'
 describe('SaveFileUseCase', () => {
         
     afterEach(() => {
-        fs.rmSync('outputs', { recursive: true })
+        const outputfolderExists = fs.existsSync('outputs')
+        if(outputfolderExists) fs.rmSync('outputs', { recursive: true })
     })
 
     test('should save file with default values', () => {
@@ -22,4 +23,25 @@ describe('SaveFileUseCase', () => {
         expect( fileContent ).toBe( options.fileContent )
 
     });
+
+    test('shpuld save file with custom values', () => {
+        
+        const saveFile = new SaveFile();
+        const options = {
+            fileContent: 'custom content',
+            fileDestionation: 'custom-ouputs/file-destination',
+            fileName: 'custom-table-name'
+        }
+
+        const filePath = `${options.fileDestionation}/${options.fileName}.txt`;
+
+        const result = saveFile.execute( options )
+        const fileExists = fs.existsSync( filePath );
+        const fileContent = fs.readFileSync( filePath, { encoding: 'utf-8'});
+
+        expect( result ).toBe( true)
+        expect( fileExists ).toBe( true)
+        expect( fileContent ).toBe( options.fileContent )
+
+    })
 });
